+++ m1_csrc/m1_virtual_kb.h
@@ -14,6 +14,7 @@
 #define M1_VIRTUAL_KB_H_
 
 uint8_t m1_vkb_get_filename(char *description, char *default_name, char *new_name);
+uint8_t m1_vkb_get_text(char *description, char *default_text, char *new_text);
 uint8_t m1_vkbs_get_data(char *description, char *new_data);
 
 #endif /* M1_VIRTUAL_KB_H_ */